add abstract class for range actions that manage ranges for all subclasses
